+++ frontend/package.json
@@ -89,6 +89,7 @@
     "postcss": "8.4.33",
     "react": "^18",
     "react-day-picker": "^8.10.1",
+    "react-diff-viewer-continued": "^3.4.0",
     "react-dom": "^18",
     "react-hook-form": "^7.55.0",
     "react-icons": "^5.5.0",
