+++ frontend/src/components/thread/tool-views/wrapper/ToolViewRegistry.tsx
